pyproject.toml
[tool.poetry]
name = "nonebot-plugin-gpt"
version = "0.1.18"
description = "Chatbot plugin based on ChatGPT for Nonebot."
authors = ["kifuan"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/kifuan/nonebot-plugin-gpt"
repository = "https://github.com/kifuan/nonebot-plugin-gpt"
include = ["font.otf"]

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-rc.1"
nonebot-adapter-onebot = "^2.1.5"
pydantic = "^1.9.0"
aiohttp = "^3.8.1"
Pillow = "^9.0.1"


[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
